hopefully this set it up properly

+++ xv6-riscv/kernel/proc.c
@@ -162,7 +162,7 @@ found:
   t->state = USED;
   t->tid = 0; 
 
-  t->trapframe->kernel_trap  = (uint64)usertrapret();
+  t->trapframe->kernel_trap  = usertrapret;
   t->trapframe->kernel_satp  = MAKE_SATP(p->pagetable);
   t->trapframe->kernel_hartid = r_tp();
   t->trapframe->kernel_sp = KSTACK(p->pid, 0) + PGSIZE;
